* commit '62d2a80913b822d93181cccd1e2f80a3ef1031e4':
  发布1.0.65-snapshot版本
  根据意见修改升级
  根据建议修改
  修改
  修改
  升级techops版本到1.0.64
  升级uniauth的多租户功能
  增加开关控制，是否强制检测是否有租户相关身份信息
  uniauth-租户化bug修复
  更新多租户相关的bug
  调整代码格式
  1 租户化uri_pattern功能 2 techops在没有权限的情况下，增加logout链接
  add ignore
  统一日志输出

+++ cas/src/main/java/com/dianrong/common/uniauth/cas/action/PersonalInfoManageAction.java
@@ -64,7 +64,7 @@ public class PersonalInfoManageAction extends AbstractAction {
 
 		// 获取用户账号
 		String account = principal.getId();
-		Integer tenancyId = (Integer)principal.getAttributes().get(CasProtocal.DianRongCas.getTenancyIdName());
+		Long tenancyId = (Long)principal.getAttributes().get(CasProtocal.DianRongCas.getTenancyIdName());
 		
 		String requestMethod = StringUtil.getObjectStr(context.getFlowScope().get(AppConstants.CAS_USERINFO_MANAGE_FLOW_REQUEST_METHOD_TYPE_KEY));
 		// 简单的以post和其他请求方式来区分去请求数据还是更新数据
@@ -84,7 +84,7 @@ public class PersonalInfoManageAction extends AbstractAction {
 	 * @return result
 	 * @throws Exception
 	 */
-	private Event queryUserInfo(final RequestContext context, String account, Integer tenancyId) throws Exception {
+	private Event queryUserInfo(final RequestContext context, String account, Long tenancyId) throws Exception {
 		UserDto userInfo = null;
 		try {
 			// 调服务获取用户信息
